renamed addContact file
